+++ packages/explorer/lib/src/i18n/messages/messages_en.dart
@@ -8,6 +8,8 @@
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
 // ignore_for_file:unused_import, file_names
+// ignore_for_file: type_annotate_public_apis
+// ignore_for_file: always_declare_return_types
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';