+++ packages/explorer/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.3.0
+### Changed
+* `uploadFiles` and `filePressed` callbacks moved from `ExplorerController` to `Explorer` widget 
+* Renamed `NavigatorExplorer -> ExplorerProvider` and `IoNavigatorExplorer -> IoExplorerProvider`
+### Added
+* More docs and some screenshots
+
 ## 1.2.0
 
 * Show context menu on mouse right button clicked